Remove unused parts of Version package & @inlinable attributes
