+++ LICENSE
@@ -12,6 +12,13 @@ furnished to do so, subject to the following conditions:
 The above copyright notice and this permission notice shall be included in all
 copies or substantial portions of the Software.
 
+This repository includes several pieces of source code authored by other parties.
+The files containing such source code are the following:
+* Sources/Version/Version.swift, which
+  is originally available at https://github.com/mxcl/Version and
+  is licensed under Apache 2.0 license (see https://github.com/mxcl/Version/blob/master/LICENSE).
+See the file headers for details.
+
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
